requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
asyncpg>=0.29.0